Don't set Ctty when execing runsc in console mode.
Libcontainer takes care of setting Ctty for us [1], so this is redundant. In
fact, the Ctty that we set in exec has been incorrect since the beginning, and
nobody has noticed.

An upcoming change in Go 1.13 [2] changes the semantics of the SysProcAttr.Ctty
field, breaking our current tests. Since we don't need to set the field
anyways, we should just delete it.

1: https://github.com/docker/libcontainer/blob/83a102cc68a09d890cce3b6c2e5c14c49e6373a0/standard_init_linux.go#L36
2: https://go-review.googlesource.com/c/go/+/178919/
PiperOrigin-RevId: 254264780